Add PoC for post previews in list template
Needs more work in terms of style and implementation, but the PoC serves
to see what it could look like.

I need to think more about how to properly split the post preview, and
make it look a little better by making the preview text not stand out as
much as the title.
Some sort of faded color of what the background is might work? Maybe
look at the Material style guidelines?